requirements.txt
fastapi==0.108.0
python-dotenv==1.0.0
uvicorn==0.25.0
# psycopg2==2.9.9
mysql-connector-python==8.2.0